requirements.txt
discord==2.2.2
openai==0.27.2
discord-py-slash-command==4.2.1
azure_cognitiveservices_speech==1.26.0
PyNaCl==1.5.0
cffi==1.15.1
cryptography==39.0.2
pycparser==2.21